% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded.R
\name{eupp_download_gridded}
\alias{eupp_download_gridded}
\title{Downloading EUPP Datasets}
\usage{
eupp_download_gridded(
  x,
  output_file,
  output_format = c("grib", "nc"),
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{x}{an object of class \code{\link{eupp_config}}.}

\item{output_file}{character length 1, name of the output file.}

\item{output_format}{character length 1, defaults to \code{"grib"} (see details).}

\item{verbose}{logical length 1, verbosity, defaults to \code{FALSE}.}

\item{overwrite}{logical length 1, defaults to \code{FALSE}. If set to \code{TRUE}
the \code{output_file} will be overwritten if needed. If \code{FALSE} and
\code{output_file} exists an error will be raised.}
}
\description{
Main function downloading data.
}
\details{
The function allows to store data sets in either GRIB version 1 or
NetCDF (classic 64bit; v3). The original data set is provided as GRIB, the 
conversion to NetCDF is done locally usning ECMWFs ecCodes tools which must
be installed when using NetCDF.

By default (\code{output_format = "guess"}) the function tires to guess the
\code{output_format} based on the \code{output_file} name and will stop if
not possible. You can always use \code{output_format} to control this explicitly.

\itemize{
   \item \code{output_format = "guess"}: Guess output format based on \code{output_file}.
   \item \code{output_format = "grb"}: Store result in GRIB version 1.
   \item \code{output_format = "nc"}: Store result as NetCDF file.
}
}
\author{
Reto Stauffer
}
