% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{eupp_config}
\alias{eupp_config}
\title{Creates EUPP Config used as Data Descriptor}
\usage{
eupp_config(
  type = c("reforecast", "forecast", "analysis"),
  kind = c("ctr", "ens", "hr"),
  level = c("surf", "pressure", "efi"),
  date,
  parameter = NULL,
  steps = NULL,
  area = NULL,
  cache = NULL,
  version = 0L
)
}
\arguments{
\item{type}{character of length \code{1} (see Details).}

\item{kind}{character of length \code{1} (see Details).}

\item{level}{character of length \code{1} (see Details).}

\item{date}{an object which can be converted to \code{POSIXct}. Character string
in ISO representation or an object of class \code{Date} or \code{POSIXt}.}

\item{parameter}{\code{NULL} (default) or character. When \code{NULL} all available
parameters will be processed. Can be used to retrieve/process subsets.}

\item{steps}{\code{NULL} (default) or integer. If set to \code{NULL} all available forecast
steps will be processed. An integer sequence can be provided to only process
specific forecast steps; given in hours (e.g., \code{c(6, 12)} for \code{+6}
and \code{+12} hour ahead forecasts).}

\item{area}{\code{NULL} (default) or an object of class \code{bbox}. Used for spatial subsetting.}

\item{cache}{\code{NULL} or character of length \code{1} pointing to an existing
directory. Is used for data caching (caching grib index information/inventories)
which can be handy to save some time.}

\item{version}{integer lengt \code{1}; version of the data set. Defaults to \code{0L}.
Typically not changed by an end-user.}
}
\value{
Returns an object of class \code{eupp_config}.
}
\description{
Generates an object of class \code{eupp_config} Containing all the required
settings for downloadin data set inventories (index files) and the data
itself.
}
\details{
Input argument \code{type}:

\itemize{
  \item \code{"reforecast"}: Reforecasts (or hindcasts).
  \item \code{"forecast"}: Forecast data.
  \item \code{"analysis"}: Analysis data; based on ECMWF ERA5 used as ground truth.
}

Input argument \code{kind}:

\itemize{
  \item \code{"ctr"}: Control run.
  \item \code{"ens"}: Ensemble members.
  \item \code{"hr"}: High-resolution forecast.
}

Input argument \code{level}:

\itemize{
  \item \code{"surf"}: Surface data.
  \item \code{"pressure"}: Pressure level data.
  \item \code{"efi"}: Extreme forecast index data.
}

Note that reforecasts always only become available on Mondays and Thursdays.
If \code{type = "reforecast"} and the \code{date} does not point to Mon/Thu
the function will throw an error.
}
\examples{
conf_A <- eupp_config(type = "reforecast", kind = "ctr", level = "surf",
                      date = "2017-01-02", parameter = "2t")

conf_B <- eupp_config(type = "reforecast", kind = "ctr", level = "surf",
                      date = "2017-01-02", parameter = "2t",
                      steps = c(6, 12))

}
\author{
Reto Stauffer
}
