% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{eupp_get_url_config}
\alias{eupp_get_url_config}
\title{Get Configuration}
\usage{
eupp_get_url_config()
}
\description{
Provides a series of parameters. Internal use only.
}
\details{
The variable \code{BASEURL} provides the ... base URL,
\code{PATTERN} is a character for the path/file pattern used to 
request data via HTTP. Not using a special format; {{variable}} will
be replaced using `gsub()` where needed. Will be extended with \code{.grb}
or \code{.grb.index} to either fetch the grib index (JSON format) or the
grib file itself.
}
\author{
Reto Stauffer
}
\keyword{internal}
