% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupp_stars.R
\name{eupp_stars}
\alias{eupp_stars}
\alias{[.eupp_stars}
\alias{[[.eupp_stars}
\alias{st_extract.eupp_stars}
\title{EUPP Stars Objects}
\usage{
\method{[}{eupp_stars}(x, ...)

\method{[[}{eupp_stars}(x, ...)

\method{st_extract}{eupp_stars}(x, at, bilinear, atname = NULL, ..., subsequent = NULL)
}
\arguments{
\item{x}{stars object}

\item{...}{currently unused.}

\item{at}{object of class \code{sf} or \code{sfc} forwarded to
\code{\link[stars]{st_extract}}}

\item{bilinear}{logical, forwarded to \code{\link[stars]{st_extract}}.
Defaults to \code{TRUE} (bilinear interpolation); \code{FALSE}
results in the use of nearest-neighbour interpolation.}

\item{atname}{character length \code{1} or \code{NULL} (default).
Can be used to specify a variable in the object \code{at} which should
be appended to the final object.}

\item{subsequent}{Defaults to \code{NULL}, please leave as it is.}
}
\value{
Returns an object of class \code{c("sf", "data.frame")}.
}
\description{
Custom class based on \code{stars} objects. The additional
class \code{eupp_stars} is used to provide some additional
methods to conveniently process the data.
}
\details{
Internally, an \code{eupp_stars} object is only a \code{stars}
object with one additional attribute (the extra class).

Depending on the stars object this function is calling itself multiple times.
This is controlled using the 'subsequent' argument which becomes a named
vector (internally). As an end-user, please leave this argument \code{NULL}.
Have not found a better solution yet.
}
\author{
Reto Stauffer

Reto Stauffer

Reto Stauffer
}
