% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded.R
\name{eupp_get_gridded}
\alias{eupp_get_gridded}
\alias{eupp_get_inventory}
\title{Getting Grib Data as stars}
\usage{
eupp_get_gridded(x, verbose = FALSE)

eupp_get_inventory(x)
}
\arguments{
\item{x}{object of class \code{\link{eupp_config}}.}

\item{verbose}{logical, sets verbosity level. Defaults to \code{FALSE}.}
}
\value{
TODO
}
\description{
Requires ncdf4 and the NetCDF Library to be installed.
}
\seealso{
\code{\link{eupp_download_gridded}}
}
\author{
Reto Stauffer
}
