% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupp_stars.R
\name{sf_append_atname_if_possible}
\alias{sf_append_atname_if_possible}
\title{Append Additional Variable to object x if possible}
\usage{
sf_append_atname_if_possible(x, at, atname)
}
\arguments{
\item{x}{sf object (interpolated data).}

\item{at}{sf object with \code{POINTS} and, possibly, a variable named \code{atname}.}

\item{atname}{character length \code{1} or \code{NULL}.}
}
\description{
Small hidden helper function used in \code{\link{st_extract.eupp_stars}}.
If the locations \code{at} contain a dedicated column called \code{name},
these names will be added to the stars object.
}
\author{
Reto Stauffer
}
