% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{eupp_config}
\alias{eupp_config}
\title{Creates EUPP Config used as Data Descriptor}
\usage{
eupp_config(
  type = c("reforecast", "forecast", "analysis"),
  kind = c("ens", "hr"),
  level = c("surf", "pressure", "efi"),
  date,
  parameter = NULL,
  steps = NULL,
  members = NULL,
  area = NULL,
  euppapi = TRUE,
  cache = NULL,
  version = 0L
)
}
\arguments{
\item{type}{character of length \code{1} (see Details).}

\item{kind}{character of length \code{1} or \code{NULL} (see Details).}

\item{level}{character of length \code{1} (see Details).}

\item{date}{an object which can be converted to \code{POSIXct}. Character string
in ISO representation or an object of class \code{Date} or \code{POSIXt}.}

\item{parameter}{\code{NULL} (default) or character. When \code{NULL} all available
parameters will be processed. Can be used to retrieve/process subsets.}

\item{steps}{\code{NULL} (default) or integer. If set to \code{NULL} all available forecast
members will be processed. An integer sequence can be provided to only process
specific forecast members.}

\item{area}{\code{NULL} (default) or an object of class \code{bbox}. Used for spatial subsetting.}

\item{euppapi}{logical, defaults to \code{TRUE}. Should the 'euppapi' be used or the
plain GRIB index files from the S3 store?}

\item{cache}{\code{NULL} or character of length \code{1} pointing to an existing
directory. Is used for data caching (caching grib index information/inventories)
which can be handy to save some time.}

\item{version}{integer lengt \code{1}; version of the data set. Defaults to \code{0L}.
Typically not changed by an end-user.}
}
\value{
Returns an object of class \code{eupp_config}.
}
\description{
Generates an object of class \code{eupp_config} Containing all the required
settings for downloadin data set inventories (index files) and the data
itself.
}
\details{
Input argument \code{type}:

\itemize{
  \item \code{"reforecast"}: Reforecasts (or hindcasts).
  \item \code{"forecast"}: Forecast data.
  \item \code{"analysis"}: Analysis data; based on ECMWF ERA5 used as ground truth.
}

Input argument \code{kind}. Will be ignored if \code{level = "efi"} (has no kind).

\itemize{
  \item \code{"ens"}: Ensemble members (allows for optional argument \code{members});
        \code{"ens"} also includes the control run (\code{member} zero).
  \item \code{"hr"}: High-resolution forecast.
}

Input argument \code{level}:

\itemize{
  \item \code{"surf"}: Surface data.
  \item \code{"pressure"}: Pressure level data.
  \item \code{"efi"}: Extreme forecast index data.
}

Note that reforecasts always only become available on Mondays and Thursdays.
If \code{type = "reforecast"} and the \code{date} does not point to Mon/Thu
the function will throw an error.

The forecast ensemble (\code{type = "forecast"}, \code{kind = "ctr"}) consists
of 50 ensemble members (\code{1-50}), the reforecasts or hindcasts
(\code{type = "forecast"}, \code{kind = "ctr"}) only consists of 20 ensemble
runs (\code{1-20}).
}
\examples{
conf_A <- eupp_config(type = "reforecast", kind = "ctr", level = "surf",
                      date = "2017-01-02", parameter = "2t")

conf_B <- eupp_config(type = "reforecast", kind = "ctr", level = "surf",
                      date = "2017-01-02", parameter = "2t",
                      steps = c(6, 12))

}
\author{
Reto Stauffer
}
