% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{eupp_config}
\alias{eupp_config}
\title{Creates EUPP Config used as Data Descriptor}
\usage{
eupp_config(
  product = c("analysis", "forecast", "reforecast"),
  level = c("surface", "pressure", "efi"),
  type = c("ens", "hr"),
  date,
  parameter = NULL,
  steps = NULL,
  members = NULL,
  area = NULL,
  cache = NULL,
  version = 0L
)
}
\arguments{
\item{product}{character of length \code{1} (see Details).}

\item{level}{character of length \code{1} or \code{NULL} (see Details).}

\item{type}{character of length \code{1} (see Details).}

\item{date}{an object which can be converted to \code{POSIXct}. Character string
in ISO representation or an object of class \code{Date} or \code{POSIXt}.}

\item{parameter}{\code{NULL} (default) or character. When \code{NULL} all available
parameters will be processed. Can be used to retrieve/process subsets.}

\item{steps}{\code{NULL} (default) or integer. If set to \code{NULL} all available forecast
steps will be processed. An integer sequence can be provided to only process
specific forecast steps; given in hours (e.g., \code{c(6, 12)} for \code{+6}
and \code{+12} hour ahead forecasts).}

\item{members}{\code{NULL} (default) or integer. If set to \code{NULL} all available forecast
members will be processed. An integer sequence can be provided to only process
specific forecast members.}

\item{area}{\code{NULL} (default) or an object of class \code{bbox}. Used for spatial subsetting.}

\item{cache}{\code{NULL} or character of length \code{1} pointing to an existing
directory. Is used for data caching (caching grib index information/inventories)
which can be handy to save some time.}

\item{version}{integer lengt \code{1}; version of the data set. Defaults to \code{0L}.
Typically not changed by an end-user.}
}
\value{
Returns an object of class \code{eupp_config}.
}
\description{
Generates an object of class \code{eupp_config} Containing all the required
settings for downloadin data set inventories (index files) and the data
itself.
}
\details{
Input argument \code{product}:
\itemize{
\item \code{"reforecast"}: Reforecasts (or hindcasts).
\item \code{"forecast"}: Forecast data.
\item \code{"analysis"}: Analysis data; based on ECMWF ERA5 used as ground truth.
}

Input argument \code{level}. Will be ignored if \code{type = "efi"} (has no level).
\itemize{
\item \code{"efi"}: Extreme forecast index (in this situation \code{type} will be ignored).
\item \code{"surf"}: Surface data.
\item \code{"pressure"}: Pressure level data.
}

Input argument \code{type}:
\itemize{
\item \code{"ens"}: Ensemble members including control run (allows for optional argument \code{members}).
\item \code{"hr"}: High-resolution forecast.
}

Note that reforecasts always only become available on Mondays and Thursdays.
If \code{type = "reforecast"} and the \code{date} does not point to Mon/Thu
the function will throw an error.
TODO(R): This is not true, this happens when processing the data.

Ensemble forecasts (\code{type = "ens"} consists of both, the ensemble control
run plus \code{50} ensemble members if \code{members} is not specified.
Same is true for reforecasts except that it only consists of \code{20} ensemble
members. Possible combinations:\tabular{llll}{
   ID \tab product \tab level \tab type \cr
   1 \tab analysis \tab surface \tab NULL \cr
   2 \tab  \tab pressure \tab NULL \cr
   11 \tab forecast \tab efi \tab  \cr
   12 \tab  \tab surface \tab ens \cr
   13 \tab  \tab  \tab hr \cr
   14 \tab  \tab pressure \tab ens \cr
   15 \tab  \tab  \tab hr \cr
   21 \tab reforecast \tab surface \tab ens \cr
   22 \tab  \tab pressure \tab ens \cr
}
}
\examples{

# Analysis
c1 <- eupp_config("analysis", "surface", date = "2017-01-01", steps = 12)
c2 <- eupp_config("analysis", "pressure", date = "2017-01-01", steps = 12)

# Forecasts
c11 <- eupp_config("forecast", "efi", date = "2017-01-01", steps = 12)
c12 <- eupp_config("forecast", "surface", "ens", date = "2017-01-01", steps = 12)
c13 <- eupp_config("forecast", "surface", "hr", date = "2017-01-01", steps = 12)
c14 <- eupp_config("forecast", "pressure", "ens", date = "2017-01-01", steps = 12)
c15 <- eupp_config("forecast", "pressure", "hr", date = "2017-01-01", steps = 12)

# Reforecasts
c21 <- eupp_config("reforecast", "surface", "ens", date = "2017-01-01", steps = 12)
c22 <- eupp_config("reforecast", "pressure", "ens", date = "2017-01-01", steps = 12)

}
\author{
Reto Stauffer
}
