% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{eupp_get_source_url}
\alias{eupp_get_source_url}
\title{Get Source URL}
\usage{
eupp_get_source_url(x, fileext = NULL, ...)
}
\arguments{
\item{x}{object of class \code{eupp_config} returned by \code{\link{eupp_config}}.}

\item{fileext}{\code{NULL} or character length \code{1}. File extension (or
URL extension) added to the URL. Used to generate URLs to GRIB index files.
Defaults to \code{NULL}.}

\item{...}{unused.}
}
\value{
Returns a character of length 1, URL to retrieve the data set.
}
\description{
For internal use only, returns the URL to the \code{.grb} file given
the input parameters. Called by the download functions.
}
\author{
Reto Stauffer
}
