% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eupp_stars.R
\name{[.eupp_stars}
\alias{[.eupp_stars}
\alias{[[.eupp_stars}
\alias{st_extract.eupp_stars}
\title{EUPP Stars Objects}
\usage{
\method{[}{eupp_stars}(x, ...)

\method{[[}{eupp_stars}(x, ...)

\method{st_extract}{eupp_stars}(x, at, bilinear = FALSE, ...)
}
\description{
Custom class based on \code{stars} objects. The additional
class \code{eupp_stars} is used to provide some additional
methods to conveniently process the data.
}
\details{
Internally, an \code{eupp_stars} object is only a \code{stars}
object with one additional attribute (the extra class).
}
\author{
Reto Stauffer

Reto Stauffer

Reto Stauffer
}
