% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded.R
\name{eupp_download_gridded}
\alias{eupp_download_gridded}
\alias{eupp_get_gridded}
\alias{eupp_get_inventory}
\title{Downloading EUPP Datasets}
\usage{
eupp_download_gridded(
  x,
  output_file,
  output_format = "grib",
  overwrite = FALSE,
  netcdf_kind = 3L,
  verbose = FALSE
)

eupp_get_gridded(x, verbose = FALSE)

eupp_get_inventory(x, times = 3L, verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{\link{eupp_config}}.}

\item{output_file}{character length 1, name of the output file.}

\item{output_format}{character length 1, defaults to \code{"grib"}.
Alternatively (somewhat experimental) \code{"nc"}.}

\item{overwrite}{logical length 1, defaults to \code{FALSE}. If set to \code{TRUE}
the \code{output_file} will be overwritten if needed. If \code{FALSE} and
\code{output_file} exists an error will be raised.}

\item{netcdf_kind}{numeric length 1, defaults to \code{3}. Controls the
\code{-k} (kind) flag when calling \code{grib_to_netcdf}. Only used
when \code{output_format = "nc"}.}

\item{verbose}{logical, sets verbosity level. Defaults to \code{FALSE}.}

\item{times}{positive numeric, defaults to \code{3L}. Number of
retries in case the GET request fails.}
}
\value{
No explicit return, invisibly returns the inventory of the file as a
\code{data.frame}.

Object of class \code{c("eupp_stars", "stars")}; see Details.
}
\description{
Main function downloading data.

Requires ncdf4 and the NetCDF Library to be installed.
}
\details{
The function allows to store data sets in either GRIB version 1 or
NetCDF (classic 64bit; v3). The original data set is provided as GRIB, the 
conversion to NetCDF is done locally usning ECMWFs ecCodes tools which must
be installed when using NetCDF.

The default mode is to download the data in the GRIB version 1 format. The data
set will be stored as \code{output_file} alongside with an \code{.rds} file containing
the information about the fields (the GRIB index information used for downloading).
Whilst this information can be read from the GRIB file (e.g,. via ecCodes \code{grib_ls}),
however, this allows us to interpolate GRIB files without the need of ecCodes to
be installed (using the \code{.rds} and \code{\link[stars]{read_stars}} via gdal).

If \code{output_format = "nc"} a request must contain only one date in \code{x}.
The reason is that NetCDF does not allow for overlapping 'time'+'step' specifications.
Somewhat inconvenient but there is no general solution for that in ecCodes (or, to 
be precise, the NetCDF standard definition). PLEASE NOTE that this is somewhat
experimental and may not always be the best choice as \code{grib_to_netcdf} will
drop unrequired dimensions, thus losing information (e.g., if only one level or only
one perturbationNumber is present).

This function interfaces \code{\link{eupp_download_gridded}}
to download the data and converts the original data set (GRIB version 1)
to NetCDF. Thus NetCDF support and ecCodes tools are required and it only
allows data for one specific date in \code{x}. See 'Details' section
in \code{link{eupp_download_gridded}}.

The function \code{\link[stars]{read_stars}} is used to read the NetCDF
file; note that the NetCDF file will be deleted after reading. The additional
class \code{eupp_stars} is used to provide additional support for processing
the data.

If \code{x} (\code{\link{eupp_config}}) contains a \code{bbox} definition
on \code{area} the data is subsetted directly.
}
\seealso{
\code{\link{eupp_download_gridded}} and \code{\link{eupp_stars}}
}
\author{
Reto Stauffer

Reto Stauffer
}
