% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded.R
\name{eupp_download_gridded}
\alias{eupp_download_gridded}
\alias{eupp_get_gridded}
\alias{eupp_get_inventory}
\title{Downloading EUPP Datasets}
\usage{
eupp_download_gridded(
  x,
  output_file,
  output_format = c("grib", "nc"),
  verbose = FALSE,
  overwrite = FALSE
)

eupp_get_gridded(x, verbose = FALSE)

eupp_get_inventory(x, verbose = FALSE)
}
\arguments{
\item{x}{object of class \code{\link{eupp_config}}.}

\item{output_file}{character length 1, name of the output file.}

\item{output_format}{character length 1 (\code{"grib"} or \code{"nc"}),
defaults to \code{"grib"} (see details).}

\item{verbose}{logical, sets verbosity level. Defaults to \code{FALSE}.}

\item{overwrite}{logical length 1, defaults to \code{FALSE}. If set to \code{TRUE}
the \code{output_file} will be overwritten if needed. If \code{FALSE} and
\code{output_file} exists an error will be raised.}
}
\value{
Object of class \code{stars}; see Details.
}
\description{
Main function downloading data.

Requires ncdf4 and the NetCDF Library to be installed.
}
\details{
The function allows to store data sets in either GRIB version 1 or
NetCDF (classic 64bit; v3). The original data set is provided as GRIB, the 
conversion to NetCDF is done locally usning ECMWFs ecCodes tools which must
be installed when using NetCDF.

This function interfaces \code{\link{eupp_download_gridded}}
to download the data and converts the original data set (GRIB version 1)
to NetCDF. Thus NetCDF support and ecCodes tools are required.

If \code{x} (\code{\link{eupp_config}}) contains a \code{bbox} definition
on \code{area} the data is subsetted directly.
}
\seealso{
\code{\link{eupp_download_gridded}}
}
\author{
Reto Stauffer

Reto Stauffer
}
